requirements.txt
flask
mysql-connector-python
pytest
requests
pytest-cov
pytest-mock